+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/sql/QueryHistoryAPI.java
@@ -7,8 +7,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * A service responsible for storing and retrieving the list of queries run against SQL endpoints,
- * serverless compute, and DLT.
+ * A service responsible for storing and retrieving the list of queries run against SQL endpoints
+ * and serverless compute.
  */
 @Generated
 public class QueryHistoryAPI {
@@ -29,7 +29,7 @@ public class QueryHistoryAPI {
   /**
    * List Queries.
    *
-   * <p>List the history of queries through SQL warehouses, serverless compute, and DLT.
+   * <p>List the history of queries through SQL warehouses, and serverless compute.
    *
    * <p>You can filter by user ID, warehouse ID, status, and time range. Most recently started
    * queries are returned first (up to max_results in request). The pagination token returned in